Refactor project component.
